+++ src/pieces/piece/knight.js
@@ -1,7 +1,7 @@
 // knight.js
-import { Piece } from "./pieces.js";
-import { highlightTile } from "../util/clickedPiece.js";
-import { indexToLocation } from "../util/findLocation.js";
+import { Piece } from "../pieces.js";
+import { highlightTile } from "../../util/clickedPiece.js";
+import { indexToLocation } from "../../util/findLocation.js";
 
 export class Knight extends Piece {
     constructor(team, startingPosition) {
@@ -125,16 +125,7 @@ export class Knight extends Piece {
             this.jumpLeftDown(chessBoard),
         ];
 
-        // Filter out null moves (moves outside the chessboard)
-        const filteredMoves = legalMoves.filter(move => move !== null);
-
-        // Highlight the valid moves on the UI
-        for (const move of filteredMoves) {
-            const chessMove = indexToLocation(move, this.team);
-            highlightTile(chessMove);
-        }
-
-        return filteredMoves;
+        return this.filterAndHighlightTiles(legalMoves, this.team);
     }
 
     renderPiece() {
