+++ src/pieces/piece/queen.js
@@ -1,6 +1,6 @@
-import { Piece } from "./pieces.js";
-import { highlightTile } from "../util/clickedPiece.js";
-import { indexToLocation } from "../util/findLocation.js";
+import { Piece } from "../pieces.js";
+import { highlightTile } from "../../util/clickedPiece.js";
+import { indexToLocation } from "../../util/findLocation.js";
 
 export class Queen extends Piece {
     constructor(team, startingPosition) {
@@ -105,16 +105,7 @@ export class Queen extends Piece {
         this.maxMove(legalMoves, 1, 1, chessBoard); // Diagonal Down-Right
         this.maxMove(legalMoves, 1, -1, chessBoard); // Diagonal Down-Left
 
-        // Filter out null moves (moves outside the chessboard)
-        const filteredMoves = legalMoves.filter(move => move !== null);
-
-        // Highlight the valid moves on the UI
-        for (const move of filteredMoves) {
-            const chessMove = indexToLocation(move, this.team);
-            highlightTile(chessMove);
-        }
-
-        return filteredMoves;
+        return this.filterAndHighlightTiles(legalMoves, this.team);
     }
 
     renderPiece() {
