+++ src/pieces/pieces.js
@@ -1,3 +1,6 @@
+import { highlightTile } from "../util/clickedPiece.js";
+import { indexToLocation } from "../util/findLocation.js";
+
 export class Piece {
     constructor(name, team, startingPosition) {
         this.name = name;
@@ -133,6 +136,18 @@ export class Piece {
                 break;
             }
         }
+    }
+
+    // Legal Moves - Highlight Tiles
+    filterAndHighlightTiles(legalMoves, piece) {
+        const filteredMoves = legalMoves.filter(move => move != null);
+
+        for (const move of filteredMoves) {
+            const chessAlgebraicMove = indexToLocation(move, piece);
+            highlightTile(chessAlgebraicMove);
+        }
+
+        return filteredMoves;
     }
 
     // Piece has been captured
