+++ src/util/checkmate/kingInCheck.js
@@ -1,0 +1,7 @@
+const enemyPotentialMoves = () => {
+    
+}
+
+const getKingLocation = () => {
+
+}
